+++ pytest_helm_charts/giantswarm_app_platform/__init__.py
@@ -1,1 +1,1 @@
-"""This package include utilities to work with Giant Swarm application platform."""
+"""This package includes utilities to work with Giant Swarm application platform."""
