+++ pytest_helm_charts/giantswarm_app_platform/api/__init__.py
@@ -1,0 +1,1 @@
+"""Kubernetes API (CRs) definitions for Giant Swarm App Platform"""
