+++ tests/test_app_platform.py
@@ -10,7 +10,7 @@ import pytest_helm_charts
 import pytest_helm_charts.api.fixtures
 import pytest_helm_charts.api.namespace
 import pytest_helm_charts.fixtures
-import pytest_helm_charts.giantswarm_app_platform.utils
+import pytest_helm_charts.giantswarm_app_platform.app
 import pytest_helm_charts.giantswarm_app_platform.fixtures
 from pytest_helm_charts.clusters import Cluster
 from pytest_helm_charts.giantswarm_app_platform.app import AppFactoryFunc
@@ -32,8 +32,8 @@ def test_app_factory_working(kube_cluster: Cluster, app_factory: AppFactoryFunc,
 
     config_values: YamlDict = {"key1": {"key2": "my-val"}}
     mocker.patch("pytest_helm_charts.giantswarm_app_platform.catalog.CatalogCR.create")
-    mocker.patch("pytest_helm_charts.giantswarm_app_platform.utils.AppCR", autospec=True)
-    mocker.patch("pytest_helm_charts.giantswarm_app_platform.utils.ConfigMap", autospec=True)
+    mocker.patch("pytest_helm_charts.giantswarm_app_platform.app.AppCR", autospec=True)
+    mocker.patch("pytest_helm_charts.giantswarm_app_platform.app.ConfigMap", autospec=True)
     mocker.patch("pytest_helm_charts.giantswarm_app_platform.app.wait_for_apps_to_run", autospec=True)
     mocker.patch("pytest_helm_charts.api.fixtures.ensure_namespace_exists", autospec=True)
     test_configured_app: ConfiguredApp = app_factory(
@@ -48,7 +48,7 @@ def test_app_factory_working(kube_cluster: Cluster, app_factory: AppFactoryFunc,
     )
 
     # assert that configMap was created for the app
-    cm = cast(unittest.mock.Mock, pytest_helm_charts.giantswarm_app_platform.utils.ConfigMap)
+    cm = cast(unittest.mock.Mock, pytest_helm_charts.giantswarm_app_platform.app.ConfigMap)
     cm.assert_called_once_with(
         kube_cluster.kube_client,
         {
@@ -66,7 +66,7 @@ def test_app_factory_working(kube_cluster: Cluster, app_factory: AppFactoryFunc,
     cast(unittest.mock.Mock, pytest_helm_charts.api.fixtures.ensure_namespace_exists).assert_called_once_with(
         kube_cluster.kube_client, app_namespace, None, None
     )
-    app_cr = cast(unittest.mock.Mock, pytest_helm_charts.giantswarm_app_platform.utils.AppCR)
+    app_cr = cast(unittest.mock.Mock, pytest_helm_charts.giantswarm_app_platform.app.AppCR)
     app_cr.assert_called_once_with(
         kube_cluster.kube_client,
         {
